ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
groq==0.11.0
langchain==0.3.1
langchain-community==0.3.1
langchain-groq==0.2.0
qdrant-client==1.11.3
sentence-transformers==3.1.1
redis==5.1.1
numpy==1.26.4
tiktoken==0.7.0
